# Optimize channel_shuffle_op on GPU