… caffe2/operators/channel_shuffle_op_gpu.cu