Summary: Optimize channel_shuffle_op on GPU

Differential Revision: D10639281