@BIT-silence 2018-10-24
Benchmark on GPU
N = 256, C = 256, H = 128, W = 128, group = 8,
NCHW: 84.82ms -> 38.84ms
NHWC: 81.83ms -> 39.01ms