@bddppq 2018-10-24
triggered c2 rocm tests run: https://ci.pytorch.org/jenkins/job/caffe2-builds/job/py2-clang7-rocmdeb-ubuntu16.04-trigger-test/3081/