@BIT-silence 2018-10-24
@pytorchbot retest this please